+++ src/demo-app/demo-material-module.ts
@@ -11,6 +11,7 @@ import {
   MdDatepickerModule,
   MdDialogModule,
   MdExpansionModule,
+  MdFormFieldModule,
   MdGridListModule,
   MdIconModule,
   MdInputModule,
@@ -50,6 +51,7 @@ import {CdkTableModule} from '@angular/cdk/table';
     MdDatepickerModule,
     MdDialogModule,
     MdExpansionModule,
+    //MdFormFieldModule,
     MdGridListModule,
     MdIconModule,
     MdInputModule,
