+++ src/lib/autocomplete/autocomplete-trigger.ts
@@ -40,7 +40,8 @@ import {Observable} from 'rxjs/Observable';
 import {MdOptionSelectionChange, MdOption} from '../core/option/option';
 import {ENTER, UP_ARROW, DOWN_ARROW, ESCAPE} from '../core/keyboard/keycodes';
 import {Directionality} from '../core/bidi/index';
-import {MdFormField, MdInput} from '../input/input-container';
+import {MdFormField} from '../form-field/index';
+import {MdInput} from '../input/input';
 import {Subscription} from 'rxjs/Subscription';
 import {merge} from 'rxjs/observable/merge';
 import {fromEvent} from 'rxjs/observable/fromEvent';
