+++ src/lib/datepicker/datepicker.spec.ts
@@ -16,6 +16,7 @@ import {
   createKeyboardEvent,
   dispatchEvent,
 } from '@angular/cdk/testing';
+import {MdFormFieldModule} from "../form-field/index";
 
 describe('MdDatepicker', () => {
   afterEach(inject([OverlayContainer], (container: OverlayContainer) => {
@@ -28,6 +29,7 @@ describe('MdDatepicker', () => {
         imports: [
           FormsModule,
           MdDatepickerModule,
+          MdFormFieldModule,
           MdInputModule,
           MdNativeDateModule,
           NoopAnimationsModule,
@@ -817,6 +819,7 @@ describe('MdDatepicker', () => {
         imports: [
           FormsModule,
           MdDatepickerModule,
+          MdFormFieldModule,
           MdInputModule,
           NoopAnimationsModule,
           ReactiveFormsModule,
@@ -840,7 +843,13 @@ describe('MdDatepicker', () => {
 
     beforeEach(async(() => {
       TestBed.configureTestingModule({
-        imports: [MdDatepickerModule, MdInputModule, MdNativeDateModule, NoopAnimationsModule],
+        imports: [
+          MdDatepickerModule,
+          MdFormFieldModule,
+          MdInputModule,
+          MdNativeDateModule,
+          NoopAnimationsModule
+        ],
         declarations: [StandardDatepicker],
       }).compileComponents();
