+++ src/lib/input/index.ts
@@ -7,43 +7,29 @@
  */
 
 import {NgModule} from '@angular/core';
-import {
-  MdError,
-  MdHint,
-  MdFormField,
-  MdInput,
-  MdPlaceholder,
-  MdPrefix,
-  MdSuffix
-} from './input-container';
+import {MdInput} from './input';
 import {MdTextareaAutosize} from './autosize';
 import {CommonModule} from '@angular/common';
 import {PlatformModule} from '../core/platform/index';
+import {MdFormFieldModule} from '../form-field/index';
 
 
 @NgModule({
   declarations: [
-    MdError,
-    MdHint,
-    MdFormField,
     MdInput,
-    MdPlaceholder,
-    MdPrefix,
-    MdSuffix,
     MdTextareaAutosize,
   ],
   imports: [
     CommonModule,
+    MdFormFieldModule,
     PlatformModule,
   ],
   exports: [
-    MdError,
-    MdHint,
-    MdFormField,
+    // TODO(mmalerba): We import and re-export the form field module since all existing users of
+    // `MdInput` will need this to continue using `md-input-container`. We may want to keep this
+    // long term since the `MdInput` directive will almost always be used with `md-form-field`.
+    MdFormFieldModule,
     MdInput,
-    MdPlaceholder,
-    MdPrefix,
-    MdSuffix,
     MdTextareaAutosize,
   ],
 })
@@ -51,6 +37,6 @@ export class MdInputModule {}
 
 
 export * from './autosize';
-export * from './input-container';
-export * from './input-container-errors';
+export * from './input';
+export * from './input-errors';
