+++ src/lib/input/input-errors.ts
@@ -1,0 +1,12 @@
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+/** @docs-private */
+export function getMdInputUnsupportedTypeError(type: string): Error {
+  return Error(`Input type "${type}" isn't supported by mdInput.`);
+}
