+++ src/lib/module.ts
@@ -48,6 +48,7 @@ import {MdExpansionModule} from './expansion/index';
 import {MdTableModule} from './table/index';
 import {MdSortModule} from './sort/index';
 import {MdPaginatorModule} from './paginator/index';
+import {MdFormFieldModule} from './form-field/index';
 
 const MATERIAL_MODULES = [
   MdAutocompleteModule,
@@ -60,6 +61,7 @@ const MATERIAL_MODULES = [
   MdTableModule,
   MdDialogModule,
   MdExpansionModule,
+  MdFormFieldModule,
   MdGridListModule,
   MdIconModule,
   MdInputModule,
