+++ src/lib/public_api.ts
@@ -25,6 +25,7 @@ export * from './checkbox/index';
 export * from './datepicker/index';
 export * from './dialog/index';
 export * from './expansion/index';
+export * from './form-field/index';
 export * from './grid-list/index';
 export * from './icon/index';
 export * from './input/index';
