+++ src/universal-app/kitchen-sink/kitchen-sink.ts
@@ -10,6 +10,7 @@ import {
   MdDatepickerModule,
   MdDialogModule,
   MdExpansionModule,
+  MdFormFieldModule,
   MdGridListModule,
   MdIconModule,
   MdInputModule,
@@ -72,6 +73,7 @@ export class KitchenSink {
     MdChipsModule,
     MdDatepickerModule,
     MdDialogModule,
+    MdFormFieldModule,
     MdGridListModule,
     MdIconModule,
     MdInputModule,
